ADD - #6 change the underlying file transferring mechanism - closes #5 because socket connection only exists when connecting via FTP
