+++ README.md
@@ -1,12 +1,20 @@
-# Ftp2Http
-Bridge for serving and downloading the FTP content via HTTP.
+# Nas2Http (was Ftp2Http)
+Personal application to serve the content on NAS via HTTP.
 
-Check [issues page][issues] for the to-be-impplemented features!
+Check [issues page][issues] for the upcoming features!
 
-[issues]: https://github.com/RaenonX/Ftp2Http/issues
+[issues]: https://github.com/RaenonX/Nas2Http/issues
 
 ### Timeline
 
 - 2020/09/09: Idea came up.
 
 - 2020/09/12: First release.
+
+- 2020/09/13: Changed the name from `Ftp2Http` to `Nas2Http` as the underlying mechanism has changed.
+
+### Note
+
+This is developed, tested and deployed on Windows.
+
+The mechanism is to forward the content on mapped network drive to the end user's client.
