+++ path.py
@@ -1,6 +1,8 @@
 """Path info object."""
 from typing import Generator, Tuple
 
+__all__ = ("PathInfo",)
+
 
 class PathInfo:
     """
@@ -8,24 +10,30 @@ class PathInfo:
 
     Returned path will always start from a slash (``/``), and end with a slash (``/``).
     """
+    ROOT = "Y:"
 
-    def __init__(self, path: str):
-        if not path.startswith("/"):
-            path = f"/{path}"
-        if not path.endswith("/"):
-            path = f"{path}/"
+    def __init__(self, path_without_root: str):
+        if not path_without_root.startswith("/"):
+            path_without_root = f"/{path_without_root}"
+        if not path_without_root.endswith("/"):
+            path_without_root = f"{path_without_root}/"
 
-        self._path = path
+        self._path = f"{PathInfo.ROOT}{path_without_root}"
+
+    @property
+    def full_path_with_root(self) -> str:
+        """Get the full path with the configured root. This always starts from ``ROOT`` and ends with "/"."""
+        return self._path
 
     @property
     def full_path(self) -> str:
-        """Get the full path. This always starts from "/" and end with "/"."""
-        return self._path
+        """Get the full path without root. This always starts from "/" and end with "/"."""
+        return self._path[len(self.ROOT):]
 
     @property
     def full_path_for_url(self) -> str:
         """This is same as ``self.full_path`` with the beginning "/" stripped off."""
-        return self._path[1:]
+        return self.full_path[1:]
 
     @property
     def full_paths_from_root(self) -> Generator[Tuple[str, str], None, None]:
@@ -47,7 +55,7 @@ class PathInfo:
         - ``/A/B/C/``
         """
         ret = ""
-        for section in self._path.split("/"):
+        for section in self.full_path.split("/"):
             if not section:
                 continue
