+++ tests/fixtures/flask.py
@@ -8,6 +8,7 @@ __all__ = ("client",)
 @pytest.fixture
 def client():
     app.config["TESTING"] = True
+    app.config["SERVER_NAME"] = "localhost"
 
-    with app.test_client() as client:
+    with app.test_client() as client, app.app_context():
         yield client
